Fix spotless check issues
